requirements.txt
# Deep learning & vision
ultralytics
opencv-python
numpy
Pillow
scipy

# Interface web
gradio

